Updated driving

+++ arduino/main/RobotController.h
@@ -55,7 +55,7 @@ private:
 	void addAction(Action::Action a);       // Set a state flag
 	void removeAction(Action::Action a);    // Unset a state flag
 	int state;                              // The current state of the robot, collection of actions defined above
-  double lastMovementUpdate;
+  unsigned long lastMovementUpdate;
 
 	// Servo wheels
 	Servo wheelLeft;
